+++ modules/home/fzf.nix
@@ -1,6 +1,9 @@
 # Fzf is a general-purpose command-line fuzzy finder.
-{ config, lib, ... }:
-let
+{
+  config,
+  lib,
+  ...
+}: let
   accent = "#" + config.lib.stylix.colors.base0D;
   foreground = "#" + config.lib.stylix.colors.base05;
   muted = "#" + config.lib.stylix.colors.base03;
@@ -25,6 +28,9 @@ in {
       "--prompt='/ '"
       "-i"
       "--no-bold"
+      "--bind='enter:execute(nvim {})'"
+      "--preview='bat --style=numbers --color=always --line-range :500 {}'"
+      "--preview-window=right:60%:wrap"
     ];
   };
 }